# UT5/Supuesto práctico proyecto/TejeraSantana_Adoney_supuesto_dsw_ut5/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supuesto UT5 Tejera Santana Adoney",
        "description": "Documentacion de supuesto ut5",
        "version": "11.31"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/adoneytj/planes_nutricionales": {
            "get": {
                "tags": [
                    "Planes_nutricionales"
                ],
                "summary": "Listar planes nutricionales",
                "description": "Lista todos los planes nutricionales, solo admin y nutricionistas tienen autorización",
                "operationId": "indexPlanesNutricionales",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o nutricionista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Planes_nutricionales"
                ],
                "summary": "Crear plan nutricional",
                "description": "Crea un plan nutricional, solo admin y nutricionistas (solo puede crear su propio plan) tienen autorización",
                "operationId": "storePlanesNutricionales",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_nutricionista",
                                    "id_cliente",
                                    "nombre",
                                    "recomendaciones_dieta",
                                    "porcentaje_carbohidratos",
                                    "porcentaje_proteinas",
                                    "porcentaje_grasas",
                                    "porcentaje_fibra",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "id_nutricionista": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hay que comer"
                                    },
                                    "recomendaciones_dieta": {
                                        "type": "string",
                                        "example": "Hay que comer mas"
                                    },
                                    "porcentaje_carbohidratos": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "porcentaje_proteinas": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "porcentaje_grasas": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "porcentaje_fibra": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-2-8"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2054-3-12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o nutricionista y que sea su propio plan"
                    },
                    "205": {
                        "description": "Algún parámetro inválido"
                    }
                }
            }
        },
        "/api/adoneytj/planes_nutricionales/{id_plan}": {
            "get": {
                "tags": [
                    "Planes_nutricionales"
                ],
                "summary": "Obtener plan nutricional",
                "description": "Obtiene un plan nutricional, solo admin y nutricionistas (solo puede con su propio plan) tienen autorización",
                "operationId": "showPlanesNutricionales",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin, nutricionista, si es su propio plan, o el cliente del plan"
                    },
                    "205": {
                        "description": "Algún parámetro inválido o no se encuentra el plan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Planes_nutricionales"
                ],
                "summary": "Actualizar plan nutricional",
                "description": "Actualiza un plan nutricional, solo admin y nutricionistas (solo puede con su propio plan) tienen autorización",
                "operationId": "updatePlanesNutricionales",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_nutricionista": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hay que comer"
                                    },
                                    "recomendaciones_dieta": {
                                        "type": "string",
                                        "example": "Hay que comer mas"
                                    },
                                    "porcentaje_carbohidratos": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "porcentaje_proteinas": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "porcentaje_grasas": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "porcentaje_fibra": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-2-8"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2054-3-12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o nutricionista y que sea su propio plan"
                    },
                    "205": {
                        "description": "Algún parámetro inválido o no se encuentra el plan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planes_nutricionales"
                ],
                "summary": "Borrar plan nutricional",
                "description": "Borra un plan nutricional, solo admin y nutricionistas (solo puede con su propio plan) tienen autorización",
                "operationId": "destroyPlanesNutricionales",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o nutricionista y que sea su propio plan"
                    },
                    "205": {
                        "description": "Algún parámetro inválido o no se encuentra el plan"
                    }
                }
            }
        },
...
    },
...
}